requirements.txt
addict==2.4.0
ansitable==0.9.10
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colored==1.4.4
colorlog==6.8.2
comm==0.2.1
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
docstring_parser==0.16
embreex==2.17.7.post4
eval-type-backport==0.1.3
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.4
Flask==3.0.3
fonttools==4.51.0
gdown==5.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.5.35
idna==3.7
imageio==2.34.0
importlib-metadata==7.0.1
importlib_resources==6.4.0
ipykernel==6.29.0
ipython==8.12.3
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lazy_loader==0.4
lxml==5.2.1
mapbox-earcut==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mdurl==0.1.2
msgpack==1.0.8
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nodeenv==1.8.0
numpy==1.24.4
open3d==0.18.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pgraph-python==0.6.2
pickleshare==0.7.5
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
plotly==5.20.0
pre-commit==3.5.0
progress==1.6
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pybullet==3.2.6
pycollada==0.8
Pygments==2.17.2
pyliblzfse==0.4.1
pyparsing==3.1.2
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
retrying==1.3.4
rich==13.7.1
roboticstoolbox-python==1.1.0
rpds-py==0.18.0
rtb-data==1.0.1
Rtree==1.2.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
shapely==2.0.3
shtab==1.7.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
spatialgeometry==1.1.0
spatialmath-python==1.1.9
stack-data==0.6.3
svg.path==6.3
swift-sim==1.1.0
tenacity==8.2.3
threadpoolctl==3.4.0
tifffile==2023.7.10
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transforms3d==0.4.1
trimesh==4.3.0
typing_extensions==4.11.0
tyro==0.8.3
tzdata==2024.1
urllib3==2.2.1
vhacdx==0.0.6
virtualenv==20.25.1
viser==0.1.28
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.2
widgetsnbextension==4.0.10
xatlas==0.0.9
xxhash==3.4.1
yourdfpy==0.0.56
zipp==3.18.1
